requirements.txt
Pillow
requests
fpdf